feat: 启用PostgreSQL支持并修复相关问题
- 启用node的PostgreSQL分支，移除临时禁用代码
- 修复PostgreSQL标识符大小写问题，统一使用引号包裹表名和列名
- 补充ASK_ORDER/BID_ORDER表的缺失字段
- 修复测试文件中的表名大小写问题
- 更新README和文档，反映PostgreSQL测试通过状态
- 完善交易处理器中的别名、投票和账户属性逻辑

+++ crates/p2p/src/daemon/discovery.rs
@@ -281,7 +281,7 @@ impl DiscoveryDaemon {
             let model = PeerModel {
                 address: addr_str,
                 last_updated: Some(now),
-                services: Some(peer.services),
+                services: Some(peer.services as i64),
             };
             if repo.upsert(&model).await.is_ok() {
                 saved += 1;
